Revert ignores CacheMap
It's premature optimization. Each .gitignore takes approx 1ms to process, so even if the target module is burried in a 6 level deep subdir, with a .gitignore file in each of those parents, it's only 6ms, and only for the same `resolveFromCaller`.

If the target module uses a custom SDK in a sibling directory 6 levels deep, all those .gitignore files will be read twice because the custom SDK will make a separate `resolveFromCaller` call.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>